refactor(client): DevQuickToggle 공용 컴포넌트로 추출
AppLayout 내부 함수 → components/DevQuickToggle.tsx 로 분리(동작 동일).
ConsoleLayout 과 공유하기 위함. AppLayout 은 공용 import 로 교체, 미사용
setDevPagesEnabled import 정리.
